demo for edl context-op
A edl-demo based on one of the models of Ji, van Rij, & Taatgen (ICCM 2019)  

This demo shows that operators are gradually associated with the positional context (of first, second, or third syllable position) vs. the contents of a-b-a instantiations (or "di", "le", "we" ...). This demonstrates the the nice discriminatory ability of edl, to learn the pecularity of the design of Marcus et al. (1999).