edl fix
Now better comprehends and revise the equation of vTotalActivation.
-mass reduction weights during the cue-outcome_absent situation are still not calculated;
-in prims, weights are used to calculate spreading, but the parameter of spreading is represented as alpha in edl, therefore needs revision still;
-Finally, "decay" vs "interference" debt. 
The "decay" theory seems to be slowly entering a disadvantage phase, but from a practical point of view, it can indeed simulate the learning-forgetting curve very well. The "interference" theory may be more realistic. But there is no massive effort of edl to model the learning-forgetting phenomena yet. In edl, mass reduction of cue-outcome_absent weights is similar to decay term. So in the option of edl, the baselevel term needs to be dropped.